// lib/api/models/AuditLogResponseDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type AuditLogResponseDto = {
    id: string;
    actor: string;
    action: string;
    entityType?: string;
    entityId?: string;
    metadata?: Record<string, any>;
    createdAt: string;
};